+++ package/src/skia/Color.ts
@@ -17,19 +17,11 @@ export const rgbaColor = (r: number, g: number, b: number, af: number) => {
 };
 
 export const processColorAsInt = (cl: number | string): SkColor => {
-  let color = Skia.Color(cl);
+  const color = Skia.Color(cl);
   if (typeof color !== "number") {
     throw new Error(`Couldn't process color: ${cl}`);
   }
-  // On android we need to move the alpha byte to the start of the structure
-  if (Platform.OS === "android") {
-    color = color >>> 0;
-    const a = (color >> 24) & 0xff;
-    const r = (color >> 16) & 0xff;
-    const g = (color >> 8) & 0xff;
-    const b = color & 0xff;
-    color = ((a << 24) | (r << 16) | (g << 8) | b) >>> 0;
-  }
+
   return color;
 };
